📚 DOCS: Update gallery.yml (#365)
I've added an entry for my book, "Jupyter Guide to Linear Algebra".  I'm still working on minor additions and editing, but the book is useful in its current form and is approaching completion.  Thank you for your work on the tools to make this possible.  Jupyter book has provided a great authoring experience for the subject.