Update gallery of books to include `HPX`
It would be a pleasure to include our project to your list of books.

`HPX` is a C++ Standard Library for Concurrency and Parallelism. It implements all of the corresponding facilities as defined by the C++ Standard. Additionally, in HPX we implement functionalities proposed as part of the ongoing C++ standardization process and we also extend the C++ Standard APIs to the distributed case.

@hkaiser